ai/pyproject.toml
[project]
name = "ajrasakha-ai"
version = "0.1.0"
description = "AI agent system for agricultural domain applications"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-mongodb>=0.2.0",
    "langchain-qwq>=0.1.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.10.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.10.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "fastmcp>=3.1.1",
    "ipykernel>=7.2.0",
    #    "sentence-transformers>=2.6.0",
    "langchain-mcp-adapters",
    "langchain-anthropic>=1.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ajrasakha"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["ajrasakha"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
